requirements.txt
transformers
torch
pillow
torchvision